// src/enums/HighwayKind.ts
enum HighwayKind {
    Motorway = "https://w3id.org/openstreetmap/terms#Motorway",
    Trunk = "https://w3id.org/openstreetmap/terms#Trunk",
    Primary = "https://w3id.org/openstreetmap/terms#Primary",
    Secondary = "https://w3id.org/openstreetmap/terms#Secondary",
    Tertiary = "https://w3id.org/openstreetmap/terms#Tertiary",
    Unclassified = "https://w3id.org/openstreetmap/terms#Unclassified",
    Residential = "https://w3id.org/openstreetmap/terms#Residential",
    MotorwayLink = "https://w3id.org/openstreetmap/terms#MotorwayLink",
    TrunkLink = "https://w3id.org/openstreetmap/terms#TrunkLink",
    PrimaryLink = "https://w3id.org/openstreetmap/terms#PrimaryLink",
    SecondaryLink = "https://w3id.org/openstreetmap/terms#SecondaryLink",
    TertiaryLink = "https://w3id.org/openstreetmap/terms#TertiaryLink",
    Service = "https://w3id.org/openstreetmap/terms#Service",
    Track = "https://w3id.org/openstreetmap/terms#Track",
    Footway = "https://w3id.org/openstreetmap/terms#Footway",
    Path = "https://w3id.org/openstreetmap/terms#Path",
    LivingStreet = "https://w3id.org/openstreetmap/terms#LivingStreet",
    Cycleway = "https://w3id.org/openstreetmap/terms#Cycleway",
}

export default HighwayKind;